feat(inventario): detalle de acta muestra formato GOR-F-084
Nuevo componente presentacional compartido acta-documento (formato oficial GOR-F-084), usado por el detalle y la vista de impresion. El detalle del acta pasa de la tarjeta estilizada al formato oficial, con boton 'Imprimir / Descargar'. Se quitan los botones 'Exportar .docx' y 'Cambiar Estado' del topbar, y las ramas muertas de firma (el acta nace firmada).
