+++ libs/inventario/inventario/src/lib/pages/actas-page/actas-detail/actas-detail.component.ts
@@ -1,17 +1,16 @@
 import {
   ChangeDetectionStrategy,
   Component,
-  computed,
   inject,
   OnInit,
 } from '@angular/core';
 import { Router, ActivatedRoute, RouterModule } from '@angular/router';
-import { UpperCasePipe } from '@angular/common';
 import {
   StatusBadgeComponent,
   LucideIconComponent,
 } from '@restaurant/shared/ui';
 import { BackButtonComponent } from '../../../components/back-button/back-button.component';
+import { ActaDocumentoComponent } from '../../../components/acta-documento/acta-documento.component';
 import { ActaEstado } from '../../../models/acta.model';
 import { ActasFacade } from '../../../data-access/actas.facade';
 
@@ -21,10 +20,10 @@ import { ActasFacade } from '../../../data-access/actas.facade';
   changeDetection: ChangeDetectionStrategy.OnPush,
   imports: [
     RouterModule,
-    UpperCasePipe,
     StatusBadgeComponent,
     LucideIconComponent,
     BackButtonComponent,
+    ActaDocumentoComponent,
   ],
   templateUrl: './actas-detail.component.html',
   styleUrl: './actas-detail.component.scss',
@@ -36,16 +35,6 @@ export class ActasDetailComponent implements OnInit {
 
   acta    = this.facade.actaSeleccionada;
   loading = this.facade.loading;
-
-  /** Nombre del instructor tomado de la lista de asistentes */
-  instructorNombre = computed(() => {
-    const a = this.acta();
-    if (!a) return a?.instructorId ?? '';
-    const firmante = a.asistentes?.find(f =>
-      f.dependenciaRol.toLowerCase().includes('instructor')
-    );
-    return firmante?.nombre || a.instructorId;
-  });
 
   ngOnInit(): void {
     const id = this.route.snapshot.paramMap.get('id');
@@ -80,48 +69,23 @@ export class ActasDetailComponent implements OnInit {
   }
 
   // ── Acciones ─────────────────────────────────────────────────────────────
-  /** Avanza al siguiente estado del flujo. Acepta un estado explícito o lo calcula automáticamente. */
-  cambiarEstado(nuevoEstado?: ActaEstado): void {
-    const id    = this.acta()?.id;
-    const estado = nuevoEstado ?? this.siguienteEstado();
-    if (id && estado) this.facade.cambiarEstado(id, estado);
-  }
-
-  private siguienteEstado(): ActaEstado | null {
-    const actual = this.acta()?.estado;
-    if (!actual) return null;
-    const flujo: Partial<Record<ActaEstado, ActaEstado>> = {
-      BORRADOR:         'PENDIENTE_FIRMAS',
-      PENDIENTE_FIRMAS: 'FIRMADA',
-      FIRMADA:          'REVISADA',
-      REVISADA:         'ARCHIVADA',
-    };
-    return flujo[actual] ?? null;
+  /** Avanza al estado indicado (p. ej. REVISADA → ARCHIVADA). */
+  cambiarEstado(nuevoEstado: ActaEstado): void {
+    const id = this.acta()?.id;
+    if (id && nuevoEstado) this.facade.cambiarEstado(id, nuevoEstado);
   }
 
   /** Avanza el acta de FIRMADA → REVISADA usando el ID del instructor como revisorId provisional. */
   revisarActa(): void {
-    const id          = this.acta()?.id;
+    const id           = this.acta()?.id;
     const instructorId = this.acta()?.instructorId ?? 'revisor-sena';
     if (id) {
       this.facade.revisarActa(id, instructorId);
     }
   }
 
-  /** POST /legalization/actas/{id}/exportar — genera el .docx vía backend. */
-  exportarDocx(): void {
-    const id = this.acta()?.id;
-    if (id) {
-      this.facade.exportarActa(id);
-    }
-  }
-
   // ── Navegación ───────────────────────────────────────────────────────────
   volver(): void {
     this.router.navigate(['/app/inventario/actas']);
-  }
-
-  cargarFirma(): void {
-    this.router.navigate(['cargar-firma'], { relativeTo: this.route });
   }
 }
