+++ libs/inventario/inventario/src/lib/pages/actas-page/actas-print/actas-print.component.html
@@ -11,176 +11,7 @@
 @if (loading()) {
   <div class="print-loading no-print">Cargando acta...</div>
 } @else if (acta(); as a) {
-
-<!-- ═══════════════════════════════════════════════════════════════
-     DOCUMENTO GOR-F-084 V02
-     ═══════════════════════════════════════════════════════════════ -->
-<div class="gor-doc">
-
-  <!-- LOGO -->
-  <div class="gor-logo-row">
-    <img src="assets/sena-logo.png" alt="SENA" class="gor-logo"
-         onerror="this.style.display='none'" />
-  </div>
-
-  <!-- TÍTULO -->
-  <table class="gor-table gor-table--full">
-    <tr>
-      <td class="gor-cell gor-cell--title" colspan="4">ACTA No. {{ a.numeroActa }}</td>
-    </tr>
-    <tr>
-      <td class="gor-cell gor-cell--header" colspan="4">
-        NOMBRE DEL COMITÉ O DE LA REUNIÓN: LEGALIZACIÓN MATERIALES DE FORMACIÓN POR SESIÓN.
-      </td>
-    </tr>
-    <tr>
-      <td class="gor-cell gor-cell--label">CIUDAD Y FECHA:</td>
-      <td class="gor-cell gor-cell--value">{{ fechaFormateada() }}</td>
-      <td class="gor-cell gor-cell--label">HORA INICIO:</td>
-      <td class="gor-cell gor-cell--value">{{ formatHora(a.horaInicio) }}</td>
-    </tr>
-    <tr>
-      <td class="gor-cell gor-cell--label" colspan="2">LUGAR Y/O ENLACE:</td>
-      <td class="gor-cell gor-cell--label">HORA FIN:</td>
-      <td class="gor-cell gor-cell--value">{{ formatHora(a.horaFin) }}</td>
-    </tr>
-    <tr>
-      <td class="gor-cell gor-cell--value" colspan="2">{{ (a.lugar ?? 'Escuela de Gastronomía') | uppercase }}</td>
-      <td class="gor-cell gor-cell--label" colspan="2">DIRECCIÓN / REGIONAL / CENTRO:</td>
-    </tr>
-    <tr>
-      <td class="gor-cell" colspan="2"></td>
-      <td class="gor-cell gor-cell--value" colspan="2">{{ (a.regional ?? 'Regional Quindío – CCYT') | uppercase }}</td>
-    </tr>
-  </table>
-
-  <!-- AGENDA -->
-  <table class="gor-table gor-table--full">
-    <tr>
-      <td class="gor-cell gor-cell--section-header" colspan="4">AGENDA O PUNTOS PARA DESARROLLAR:</td>
-    </tr>
-    <tr>
-      <td class="gor-cell gor-cell--text" colspan="4">
-        1. Descripción sesión de formación.<br>
-        2. Listado de materiales utilizados.
-      </td>
-    </tr>
-    <tr>
-      <td class="gor-cell gor-cell--section-header" colspan="4">OBJETIVO(S) DE LA REUNIÓN:</td>
-    </tr>
-    <tr>
-      <td class="gor-cell gor-cell--text" colspan="4">
-        1. Legalizar la entrega de materiales por sesión de formación a los aprendices y su uso de acuerdo con la planeación pedagógica del instructor.
-      </td>
-    </tr>
-  </table>
-
-  <!-- DESARROLLO -->
-  <table class="gor-table gor-table--full">
-    <tr>
-      <td class="gor-cell gor-cell--section-center">DESARROLLO DE LA REUNIÓN</td>
-    </tr>
-    <tr>
-      <td class="gor-cell gor-cell--text gor-cell--develop">
-        Durante la sesión de formación realizada durante la fecha y horario de esta acta, precedida
-        por el instructor <strong>{{ instructorNombre() }}</strong>, programado en el resultado de
-        aprendizaje: {{ a.resultadoAprendizaje ?? '—' }}, con la ficha: {{ a.fichaId }}.
-        <br><br>
-        Se realizan las siguientes actividades: {{ a.actividadesRealizadas ?? '—' }}.
-        <br><br>
-        2. Para cumplir con los objetivos de la formación, se utilizaron los materiales de formación
-        listados en la requisición de insumos diario número 45-S, adjunta a esta acta.
-      </td>
-    </tr>
-  </table>
-
-  <!-- CONCLUSIONES -->
-  <table class="gor-table gor-table--full">
-    <tr>
-      <td class="gor-cell gor-cell--section-center">CONCLUSIONES</td>
-    </tr>
-    <tr>
-      <td class="gor-cell gor-cell--text">
-        Los materiales de formación utilizados fueron debidamente legalizados por el instructor y se
-        deja como evidencia de ello: acta presente, requisición de insumos diario número 45-S
-        firmada por el vocero del grupo y registro de asistencia de aprendices de la ficha.
-      </td>
-    </tr>
-  </table>
-
-  <!-- COMPROMISOS -->
-  <table class="gor-table gor-table--full">
-    <tr>
-      <td class="gor-cell gor-cell--section-center" colspan="4">ESTABLECIMIENTO Y ACEPTACIÓN DE COMPROMISOS</td>
-    </tr>
-    <tr>
-      <td class="gor-cell gor-cell--col-header">ACTIVIDAD /DECISIÓN</td>
-      <td class="gor-cell gor-cell--col-header">FECHA</td>
-      <td class="gor-cell gor-cell--col-header">RESPONSABLE</td>
-      <td class="gor-cell gor-cell--col-header">FIRMA O PARTICIPACIÓN VIRTUAL</td>
-    </tr>
-    @for (c of (a.compromisos ?? []); track c.actividad) {
-      <tr>
-        <td class="gor-cell gor-cell--text">{{ c.actividad }}</td>
-        <td class="gor-cell gor-cell--text gor-cell--center">{{ c.fechaLimite }}</td>
-        <td class="gor-cell gor-cell--text">{{ c.responsable | uppercase }}</td>
-        <td class="gor-cell"></td>
-      </tr>
-    }
-    @if ((a.compromisos ?? []).length === 0) {
-      <tr><td class="gor-cell gor-cell--text gor-cell--center" colspan="4">—</td></tr>
-    }
-  </table>
-
-  <!-- ASISTENTES -->
-  <table class="gor-table gor-table--full">
-    <tr>
-      <td class="gor-cell gor-cell--section-center" colspan="5">DE: ASISTENTES Y APROBACIÓN DECISIONES</td>
-    </tr>
-    <tr>
-      <td class="gor-cell gor-cell--col-header">NOMBRE</td>
-      <td class="gor-cell gor-cell--col-header">DEPENDENCIA / EMPRESA</td>
-      <td class="gor-cell gor-cell--col-header">APRUEBA (SI/NO)</td>
-      <td class="gor-cell gor-cell--col-header">OBSERVACIÓN</td>
-      <td class="gor-cell gor-cell--col-header">FIRMA O PARTICIPACIÓN VIRTUAL</td>
-    </tr>
-    @for (firmante of (a.asistentes ?? []); track firmante.nombre) {
-      <tr>
-        <td class="gor-cell gor-cell--text gor-cell--center">{{ firmante.nombre | uppercase }}</td>
-        <td class="gor-cell gor-cell--text gor-cell--center">{{ firmante.dependenciaRol | uppercase }}</td>
-        <td class="gor-cell gor-cell--text gor-cell--center">{{ firmante.aprueba ? 'SI' : 'NO' }}</td>
-        <td class="gor-cell gor-cell--text gor-cell--center">-</td>
-        <td class="gor-cell"></td>
-      </tr>
-    }
-    @if ((a.asistentes ?? []).length === 0) {
-      <tr><td class="gor-cell gor-cell--text gor-cell--center" colspan="5">Sin asistentes registrados</td></tr>
-    }
-  </table>
-
-  <!-- CLAUSULA DE DATOS -->
-  <div class="gor-clausula">
-    De acuerdo con La Ley 1581 de 2012, Protección de Datos Personales, el Servicio Nacional de
-    Aprendizaje SENA, se compromete a garantizar la seguridad y protección de los datos personales
-    que se encuentran almacenados en este documento, y les dará el tratamiento correspondiente en
-    cumplimiento de lo establecido legalmente.
-  </div>
-
-  <!-- ANEXOS -->
-  <table class="gor-table gor-table--full">
-    <tr>
-      <td class="gor-cell gor-cell--section-center">ANEXOS</td>
-    </tr>
-    <tr>
-      <td class="gor-cell gor-cell--anexos"></td>
-    </tr>
-  </table>
-
-  <!-- PIE DE PÁGINA -->
-  <div class="gor-footer">GOR-F-084 V02</div>
-
-</div>
-
+  <inventario-acta-documento [acta]="a" />
 } @else {
   <div class="print-loading no-print">Acta no encontrada.</div>
 }
